pyproject.toml
[project]
name = "halfmarathon"
version = "0.0.1"
description = "Comparison of long-running AI agent approaches"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [{name = "Prachit Bhike"}]
dependencies = [
    "pydantic>=2.7",
    "anthropic>=0.62",
]

[project.optional-dependencies]
langgraph = [
    "langgraph>=1.1",
    "langgraph-checkpoint-sqlite>=3.0",
]
temporal-pydantic = [
    "pydantic-ai-slim[anthropic,temporal]>=1.86",
    "temporalio>=1.26",
]
letta = [
    "letta-client>=1.10",
]
claude-sdk = [
    "claude-agent-sdk>=0.1.65",
]
eval = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "rich>=13.7",
]
dev = [
    "halfmarathon[eval,langgraph,claude-sdk,temporal-pydantic,letta]",
    "ruff>=0.5",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = ["PLR0913", "PLR2004"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["task", "eval"]
